Fix buffer overflow in polynomial packing
In the original code, we would pack NTRU_N elements, which
would be prime and thus always require a partial last elements.
After the optimizations from eea185b, we pack NTRU_N-1 elements,
which may be a multiple of 5 (i.e. the ternary elements per byte).
In particular, this is the case for N=701.