Fix trailing bits of ciphertext test
h/t Oussama Danba. The cmov function expects a 0/1 flag, but
the trailing bits of ciphertext test was outputting other non-zero
values on failure. This caused the 'implicit rejection key' to be
set to a mixture of the 'implicit rejection key' and the 'success key'.

We did not identify a security risk from this, although it is certainly
not the correct behavior.